refactor(pci): move regions list to components
