+++ packages/manager/modules/pci/src/components/project/regions-list/regions-list.html
