+++ packages/manager/modules/pci/src/components/project/regions-list/translations/Messages_fr_FR.json
